[refactor] #125 목표 컨트롤러 병합
