+++ src/main/java/com/book/backend/domain/goal/controller/GoalController.java
@@ -2,9 +2,11 @@ package com.book.backend.domain.goal.controller;
 
 import com.book.backend.domain.goal.dto.GoalDto;
 import com.book.backend.domain.goal.dto.RecordIntervalDto;
+import com.book.backend.domain.goal.service.GoalRecordsService;
 import com.book.backend.domain.goal.service.GoalService;
 import com.book.backend.domain.openapi.service.RequestValidate;
 import com.book.backend.global.ResponseTemplate;
+import io.lettuce.core.dynamic.annotation.Param;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.media.Content;
@@ -23,13 +25,15 @@ import java.util.List;
 @RequestMapping("/api/goal")
 @RequiredArgsConstructor
 @Slf4j
-@Tag(name = "목표", description = "목표 삭제 / 목표 조회 / 목표 생성 / 목표 완료")
+@Tag(name = "목표", description = "목표 삭제 / 목표 조회 / 유저 목표 조회 / 전체 목표에 대한 합산 일주일 기록 조회 " +
+        "/ 목표 생성 / 기록 추가 / 목표 완료")
 public class GoalController {
     private final GoalService goalService;
+    private final GoalRecordsService goalRecordsService;
     private final RequestValidate requestValidate;
     private final ResponseTemplate responseTemplate;
 
-    @Operation(summary = "목표 정보 조회", description = "목표 ID에 해당하는 목표의 정보를 조회합니다.",
+    @Operation(summary = "목표 조회", description = "목표 ID에 해당하는 목표의 정보를 조회합니다.",
             parameters = {
                     @Parameter(name = "goalId", description = "목표 ID")
             },
@@ -63,7 +67,7 @@ public class GoalController {
     public ResponseEntity<?> getTotalAWeekRecords() {
         log.trace("GoalController > getTotalAWeekRecords()");
 
-        List<RecordIntervalDto> totalAWeekRecords = goalService.getTotalAWeekRecords();
+        List<RecordIntervalDto> totalAWeekRecords = goalRecordsService.getTotalAWeekRecords();
 
         return responseTemplate.success(totalAWeekRecords, HttpStatus.OK);
     }
@@ -108,12 +112,31 @@ public class GoalController {
                     @Parameter(name = "goalId", description = "목표 ID")
             },
             responses = {@ApiResponse(responseCode = "200")})
-    @DeleteMapping("delete")
+    @DeleteMapping("/delete")
     public ResponseEntity<?> deleteGoal(@RequestParam Long goalId) {
         log.trace("GoalController > deleteGoal()");
 
         goalService.deleteGoal(goalId);
 
         return responseTemplate.success("목표가 성공적으로 삭제되었습니다.", HttpStatus.OK);
+    }
+
+    @Operation(summary = "기록 추가", description = "목표 ID에 해당하는 목표에 기록을 추가합니다. " +
+            "최근 기록에서 중단한 페이지보다 큰 페이지가 입력되어야 합니다.",
+            parameters = {
+                    @Parameter(name = "goalId", description = "목표 ID"),
+                    @Parameter(name = "recentPage", description = "중단한 페이지")
+            },
+            responses = {@ApiResponse(responseCode = "200", content = @Content(schema = @Schema(implementation = GoalDto.class)),
+                    description = GoalDto.description)})
+    @PostMapping("/addRecord")
+    public ResponseEntity<?> addRecord(@RequestParam Long goalId,
+                                       @RequestParam Integer recentPage) throws Exception {
+        log.trace("GoalController > addRecord()");
+        requestValidate.isValidBookPageNum(recentPage);
+
+        GoalDto goalDto = goalRecordsService.addRecord(goalId, recentPage);
+
+        return responseTemplate.success(goalDto, HttpStatus.OK);
     }
 }
