+++ src/main/java/com/book/backend/domain/goal/service/GoalRecordsService.java
@@ -2,6 +2,7 @@ package com.book.backend.domain.goal.service;
 
 import com.book.backend.domain.goal.dto.GoalDto;
 import com.book.backend.domain.goal.dto.RecordDto;
+import com.book.backend.domain.goal.dto.RecordIntervalDto;
 import com.book.backend.domain.goal.entity.Goal;
 import com.book.backend.domain.goal.mapper.GoalMapper;
 import com.book.backend.domain.goal.repository.GoalRepository;
@@ -78,6 +79,40 @@ public class GoalRecordsService {
         goalRepository.save(goal);
 
         return goalMapper.convertToGoalDto(goal);
+    }
+
+    public List<RecordIntervalDto> getTotalAWeekRecords() {
+        log.trace("GoalService > getTotalAWeekRecords()");
+
+        // 유저 검증
+        User user = goalRequestValidate.validateAndGetLoggedInUser();
+        List<RecordIntervalDto> totalAWeekRecords = goalMapper.initializesAWeekRecords();
+
+        List<Goal> goals = user.getGoals();
+        for (Goal goal : goals) {
+            List<RecordDto> records = goal.getRecords();
+            List<RecordIntervalDto> aWeekRecords = goalMapper.convertAWeekRecords(records);
+
+            // 날짜별로 합산 처리
+            for (RecordIntervalDto record : aWeekRecords) {
+                LocalDate recordDate = record.getDate();
+                // userAWeekRecords에서 동일한 날짜를 찾아 pageInterval 합산
+                for (RecordIntervalDto totalRecord : totalAWeekRecords) {
+                    if (totalRecord.getDate().equals(recordDate)) {
+                        // 기존 값에 새로운 값 합산, null이면 0으로 처리
+                        Integer existingInterval = totalRecord.getPageInterval();
+                        Integer newInterval = record.getPageInterval();
+                        totalRecord.setPageInterval(
+                                (existingInterval == null ? 0 : existingInterval)
+                                        + (newInterval == null ? 0 : newInterval)
+                        );
+                        break;  // 해당 날짜에 대한 처리 완료 후 반복 중단
+                    }
+                }
+            }
+        }
+
+        return totalAWeekRecords;
     }
 
     private RecordDto getMostRecentRecord(Goal goal) {
